+++ src/kiara_modules/core/table/query.py
@@ -165,7 +165,7 @@ class QueryTableSQLModuleConfig(ModuleTypeConfigSchema):
     )
     relation_name: typing.Optional[str] = Field(
         description="The name the table is referred to in the sql query. If not specified, the user will be able to provide their own.",
-        default=None,
+        default="data",
     )
 
 
@@ -190,8 +190,6 @@ class QueryTableSQL(KiaraModule):
 
         if self.get_config_value("query") is None:
             inputs["query"] = {"type": "string", "doc": "The query."}
-
-        if self.get_config_value("relation_name") is None:
             inputs["relation_name"] = {
                 "type": "string",
                 "doc": "The name the table is referred to in the sql query.",
@@ -212,18 +210,19 @@ class QueryTableSQL(KiaraModule):
 
         import duckdb
 
-        _relation_name: str = self.get_config_value("relation_name")
-        if not _relation_name:
-            _relation_name = inputs.get_value_data("relation_name")
+        if self.get_config_value("query") is None:
+            _query: str = inputs.get_value_data("query")
+            _relation_name: str = inputs.get_value_data("relation_name")
+        else:
+            _query = self.get_config_value("query")
+            _relation_name = self.get_config_value("relation_name")
+
         if _relation_name.upper() in RESERVED_SQL_KEYWORDS:
             raise KiaraProcessingException(
                 f"Invalid relation name '{_relation_name}': this is a reserved sql keyword, please select a different name."
             )
 
         _table = inputs.get_value_data("table")
-        _query: str = self.get_config_value("query")
-        if not _query:
-            _query = inputs.get_value_data("query")
 
         relation: duckdb.DuckDBPyRelation = duckdb.arrow(_table)
         result: duckdb.DuckDBPyResult = relation.query(_relation_name, _query)